Fix static_cast that triggers compile errors with -Wpedantic (#1632)
(cherry picked from commit 645dd5485b2c35ea2a5a7b3433c838585e6c5972)

# Conflicts:
#	src/Camera.cc
#	src/ParticleEmitter.cc
#	src/Sensor.cc

+++ src/Sensor.cc
@@ -611,7 +611,7 @@ void Sensor::SetUpdateRate(double _hz)
 /////////////////////////////////////////////////
 std::string Sensor::TypeStr() const
 {
-  size_t index = static_cast<int>(this->dataPtr->type);
+  size_t index = static_cast<size_t>(this->dataPtr->type);
   if (index > 0 && index < sensorTypeStrs.size())
     return sensorTypeStrs[index];
   return "none";
